backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
PyJWT==2.12.1
PyMySQL==1.1.1
psycopg[binary]==3.2.9
